Общая задача коммита - обеспечить работу авторизованных пользователей только со своими задачами.
1. Добавлен столбец task.user_id как bigint, not null; на него наложено ограничение foreign key со ссылкой на user.id; по user_id создан индекс;
2. В сущность Task добавлено поле long user_id;
3. TaskController:
  3.1. currentUserId() [добавлен] - возвращает id текущего пользователя;
  3.2. create() [изменён] - создаваемой задаче прописывается id текущего пользователя;
  3.3. getAll() [изменён] - теперь обращается к методу findAllByUser_id();
  3.4. findById() [изменён] - теперь обращается к методу findByIdAndUser_id();
  3.5. checkUserIdByTaskId() [добавлен] - принимает id задачи, сравнивает user_id задачи с id текущего пользователя;
  3.6. update() [изменён] - обращается к checkUserIdByTaskId();
  3.7. delete() [изменён] - обращается к checkUserIdByTaskId();
  3.8. setDone() [изменён] - обращается к checkUserIdByTaskId();
  3.9. setUndone() [добавлен] - отмечает задачу как не выполненную;
4. UserController:
  4.1. create() - пароль возвращаемого пользователя заменяется на null;
  4.2. getCurrent() - теперь запрос выглядит как "/me", а не "/users/me";
5. SecurityConfig: configure() - при выходе из аккаунта возвращается только статус 200 OK;
6. TaskRepository:
  6.1. findAllByUser_id() - возвращает все задачи текущего пользователя;
  6.2. findByIdAndUser_id() - ищет задачу по id и user_id;
  6.3. markAsUndone() - ищет задачу по id и устанавливает значение done равным false.

+++ src/main/java/com/example/demo/controller/TaskController.java
@@ -2,36 +2,37 @@ package com.example.demo.controller;
 
 import com.example.demo.model.Task;
 import com.example.demo.repository.TaskRepository;
+import com.example.demo.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.Optional;
 
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/tasks")
 public class TaskController {
     private final TaskRepository taskRepository;
+    private final UserService userService;
 
     @RequestMapping(method = RequestMethod.POST)
     public Task create(@RequestBody Task task) {
+        task.setUser_id(currentUserId());
         return taskRepository.save(task);
     }
 
     @RequestMapping(method = RequestMethod.GET)
     public Iterable<Task> getAll() {
-        return taskRepository.findAll();
+        return taskRepository.findAllByUser_id(currentUserId());
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
     public Task findById(@PathVariable long id) {
-        return taskRepository.findById(id).orElse(null);
+        return taskRepository.findByIdAndUser_id(id, currentUserId());
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.PUT)
     public Task update(@PathVariable long id, @RequestBody Task task) {
         Task prevtask = taskRepository.findById(id).orElse(null);
-        if (prevtask != null) {
+        if (prevtask != null && checkUserIdByTaskId(id)) {
             if (task.getDate() != null) {
                 prevtask.setDate(task.getDate());
             }
@@ -39,18 +40,40 @@ public class TaskController {
                 prevtask.setDescription(task.getDescription());
             }
             return taskRepository.save(prevtask);
-        } else {
-            return null;
         }
+        return null;
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
     public void delete(@PathVariable long id) {
-        taskRepository.deleteById(id);
+        if (checkUserIdByTaskId(id)) {
+            taskRepository.deleteById(id);
+        }
     }
 
     @RequestMapping(value = "/{id}:mark-as-done", method = RequestMethod.PATCH)
     public void setDone(@PathVariable Long id) {
-        taskRepository.markAsDone(id);
+        if (checkUserIdByTaskId(id)) {
+            taskRepository.markAsDone(id);
+        }
+    }
+
+    @RequestMapping(value = "/{id}:mark-as-undone", method = RequestMethod.PATCH)
+    public void setUndone(@PathVariable Long id) {
+        if (checkUserIdByTaskId(id)) {
+            taskRepository.markAsUndone(id);
+        }
+    }
+
+    private long currentUserId() {
+        return userService.getCurrentUser().getId();
+    }
+
+    private boolean checkUserIdByTaskId(long id) {
+        Task task = taskRepository.findById(id).orElse(null);
+        if (task != null) {
+            return currentUserId() == task.getUser_id();
+        }
+        return false;
     }
 }
